Move sockets to utils/handles
Move sockets.c and sockets.h to utils/handles and
the change the including path in the file that includes
sockets.h
