+++ cclient/http/http.c
@@ -9,7 +9,7 @@
 #include <string.h>
 #include "cclient/service.h"
 #include "http_parser.h"
-#include "network/sockets.h"
+#include "utils/handles/sockets.h"
 #include "utils/macros.h"
 
 typedef enum {
