--- network/BUILD
@@ -1,7 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "sockets",
-    srcs = ["sockets.c"],
-    hdrs = ["sockets.h"],
-)
